api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
psycopg = {extras = ["binary"], version = "^3.1.13"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
